2.2.8 2/3/2019 Remove \n in the last title of columns.
