+++ Core/Ezfx.Csv/Core/CsvColumn.cs
@@ -13,14 +13,23 @@ namespace Ezfx.Csv
 
         public CsvColumn(DataColumn col)
         {
-            this.Name = col.ColumnName;
+            this.Name = col.ColumnName.Trim();
             this.Ordinal = col.Ordinal;
             this.Alias = string.Empty;
             this.Description = string.Empty;
             this.IsRequired = false;
         }
 
-        public string Name { get; set; }
+        private string _name = string.Empty;
+        public string Name { get
+            {
+                return _name;
+            }
+            set {
+                _name = value.Trim();
+            }
+        }
+
         public int Ordinal { get; set; }
         public string Alias { get; set; }
