+++ Samples/LanguageConfigSample/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.16100")]
-[assembly: AssemblyFileVersion("1.0.0.16100")]
+[assembly: AssemblyVersion("2.2.8.203")]
+[assembly: AssemblyFileVersion("2.2.8.203")]
