+++ Templates/Ezfx.Csv.Templates.CsvClass.CS/CsvClass.CSharp.vstemplate
@@ -21,7 +21,7 @@
     <ProjectItem ReplaceParameters="true">Csv.cs</ProjectItem>
   </TemplateContent>
   <WizardExtension>
-    <Assembly>Ezfx.Csv.ItemWizards, Version=1.0.0.16100, Culture=Neutral, PublicKeyToken=f6bc0ca2053018a4</Assembly>
+    <Assembly>Ezfx.Csv.ItemWizards, Version=2.2.8.203, Culture=Neutral, PublicKeyToken=f6bc0ca2053018a4</Assembly>
     <FullClassName>Ezfx.Csv.ItemWizards.CSClassWizard</FullClassName>
   </WizardExtension>
 </VSTemplate>
