+++ Templates/Ezfx.Csv.Templates/Ezfx.Csv.ItemWizards.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Wizard\CsvColumnExtensions.cs" />
     <Compile Include="Wizard\CSClassWizard.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Wizard\EncodingComboboxItem.cs" />
     <Compile Include="Wizard\VBClassWizard.cs" />
   </ItemGroup>
   <ItemGroup>
