+++ Templates/Ezfx.Csv.Templates/Wizard/ConfigForm.cs
@@ -7,26 +7,7 @@ using System.Windows.Forms;
 
 namespace Ezfx.Csv.ItemWizards
 {
-    public class EncodingComboboxItem
-    {
-        public EncodingInfo Value { get; set; }
-        public string Text
-        {
-            get { return Value.DisplayName; }
-        }
-
-        public int CodePage
-        {
-            get { return Value.CodePage; }
-        }
-
-        public override string ToString()
-        {
-            return Value.DisplayName;
-        }
-    }
-
-
+    
     public partial class ConfigForm : Form
     {
