+++ Templates/Ezfx.Csv2Entity.VSIX/source.extension.vsixmanifest
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
     <Metadata>
-        <Identity Id="Ezfx.Csv2Entity.VSIX.1d68cd43-eda3-416b-8d80-ce055c8d160a" Version="2.2.0.16500" Language="en-US" Publisher="Eric Zhou" />
+        <Identity Id="Ezfx.Csv2Entity.VSIX.1d68cd43-eda3-416b-8d80-ce055c8d1224" Version="2.2.8.203" Language="en-US" Publisher="Eric Zhou" />
         <DisplayName>Ezfx.Csv2Entity.VSIX</DisplayName>
         <Description xml:space="preserve">Used like Entity Framework with CSV. Create C# or VB.net classes automatically based on the csv file. </Description>
         <MoreInfo>https://github.com/juwikuang/csv2entity</MoreInfo>
